Issue #TG-1017 feat: Refactor DruidQueryProcessor
